[REFACTOR] Rename pyprot.protein -> pyprot.base
