+++ pyprot/align/score.py
@@ -1,6 +1,6 @@
 from math import sqrt, log, ceil
 
-from pyprot.protein import AminoAcid, Sequence, loadFasta
+from pyprot.base import AminoAcid, Sequence, loadFasta
 
 
 class ScoreMatrix:
